修改了一些逻辑错误，已经可以正常运行

+++ deps/random.hpp
@@ -4,7 +4,7 @@
 
 class Random {
 public:
-    static int uniform(int max = 1) {
-        return (std::rand() / RAND_MAX) * max;
+    static int uniform(double max = 1) {
+        return ((double)std::rand() / RAND_MAX) * max;
     }
 };
